13 Hub Location - "Add crafting functionality to the Workshop room
WorkshopRoom.cs file was deleted and then recreated in a new subfolder with significant changes. The changes include a list of craftable items (_craftableItems) and methods to populate a dropdown with these items. There was also the function to craft an item if enough resources are present, decreasing the resources accordantly.

singletonss debug logging replaced with ConvenientLogger.Log in multiple files. This change was made to improve the modularity and control over debugging prints.

Several files in both Asset and Script folders have been updated to incorporate these changes. A new CraftableItem class was also added to manage the items that can be crafted within the Workshop room."