13 Hub Location - "Added chargable property to actions
A new property `IsChargable` added to the actions parameters. This bool indicates if an action has at least one chargable point. It enhances the performance as the chargeable status of an action no longer needs to be computed based on the number of charges.
The new property 'IsChargeable' has been visualized on the UI. The 'ChargesLeft' component has been added to display the remaining charges on the action button.