13 Hub Location - Added resource craft properties and related scriptable object
Added two new files for managing different properties of craftable resources in the Hub Location. The first file, ResourceCraftPropertySO.cs, contains a scriptable object that acts as a dictionary of all craftable resources along with their properties inside the game. The OnValidate method ensures the proper assigning of resource names and types. The second file, ResourceCraftProperty.cs, creates a simple structure that holds the name, type, and cost of a craftable resource. These changes are needed in order to manage and store the various properties related to each craftable resource.