13 Hub Location - Add Quantity property to ResourceCrafting and new ResourceTypes
In ResourceCraftPropertySO.asset, added 'Quantity' field to 'ParalyzingArrows' and 'ExplosionPotion' and added a new resource 'Gold' with its 'ResourceType', 'Quantity' and 'Cost'. Also updated corresponding scripts to reflect this change. This was necessary to ensure the correct amount of resources are increased after crafting and to expand the types of resources in the game. 'Gold' was added as a new resource along with the necessary adjustments in the associated scripts.