16_save_system - PoC for new refactored version Refactored save system with persistence capability to improve efficiency. This change allows certain data to persist through instances therefore reducing the amount of duplicate data.
Changes implemented include the deletion of IDataPersistence.cs due to restructuring of save functionality, the addition of ISaveable.cs interface to standardize capabilities, the modification of data schemas and several scripts modularization and housekeeping changes for neatness.

Changes also include the adjustment of the file storage configuration to improve readability and the improvement of several safety checks to prevent null exceptions.

The refactoring took place to improve scalability and streamline loading and saving processes. The efforts have made future expansions and adjustment much easier and uniform.