16 Save system - Added save functionality in main menu
A GameObject with a button and corresponding Text UI was added in the MainMenuScene.unity. The button OnClick event triggers the "SaveGame" function in the 'DataPersistenceManager.cs' script, which was set from private to public to make it accessible. For consistency, the autosave functionality in 'DataPersistanceManager.prefab' was deactivated. This provides the player with the option to manually save the game through the main menu, rather than autosaving.