16 Save system - "Enabled autosave functionality in main menu"
In this commit, the autosave functionality was enabled in the main menu scene. This was implemented by activating the "_isAutosaveActive" flag in the DataPersistenceManager scripts and prefab. Removing the autosave action from "OnApplicationQuit" method emphasizes the input from the main menu scene. The change allows the game to save automatically, improving user experience by safeguarding their progress.